renaming everything and also better consistency
