+++ frontend/app/components/profile/profileTabs.tsx
@@ -44,7 +44,7 @@ const AttacksArray = ({attackData}: {attackData: AttackSchema[] | undefined}) =>
             attackData?.map((attack) => {
                 console.log(attackData)
                 return <Link to={`/attack/${attack.attack_id}`} className="flex flex-col items-center">
-                    <ImageWithLoader src={attack.image} className="w-40 h-40 object-cover" 
+                    <ImageWithLoader src={attack.image_link} className="w-40 h-40 object-cover" 
                     spoiler = {attack?.warnings}/>
                     <div className="w-40 text-center text-ellipsis overflow-hidden">{attack.title}</div>
