+++ frontend/app/pages/character/characterProfile.tsx
@@ -18,7 +18,7 @@ export default function CharacterPage() {
 
     const [charaData, charaLoading, charaError] = getCharacterHook(params.characterId)
 
-    const [userData, userLoading, userError] = getUserInfoHook(charaData?.owner)
+    const [userData, userLoading, userError] = getUserInfoHook(charaData?.owner_id)
 
     const [attackData, attackLoading, attackError] = getDefencesByCharacterHook(params.characterId)
 
@@ -51,7 +51,7 @@ export default function CharacterPage() {
             attackData?.map((attack) => {
                 console.log(attackData)
                 return <Link to={`/attack/${attack.attack_id}`} className="flex flex-col items-center">
-                    <ImageWithLoader src={attack.image} className="w-40 h-40 object-cover" 
+                    <ImageWithLoader src={attack.image_link} className="w-40 h-40 object-cover" 
                     spoiler = {attack?.warnings}/>
                     <div className="w-40 text-center text-ellipsis overflow-hidden">{attack.title}</div>
