+++ frontend/app/pages/login.tsx
@@ -1,7 +1,7 @@
 import { useForm, type SubmitHandler } from "react-hook-form"
 import { type UseFormRegister } from "react-hook-form"
 import { toast } from "react-toastify"
-import { signIn } from "~/api/firebase"
+import { signIn } from "~/api/supabase"
 import { useNavigate } from "react-router"
 import { useState } from "react"
 import { handleToast } from "~/functions/handleToast"
