+++ frontend/app/provider/profileProvider.tsx
@@ -1,7 +1,7 @@
 import { useState, useContext, createContext } from 'react';
 import { useEffect } from "react";
 import { onAuthStateChanged, type User } from "firebase/auth";
-import { signIn, auth, logOut } from "~/api/firebase"
+import { signIn, auth, logOut } from "~/api/supabase"
 import { getUserInfo, getUserInfoHook } from "~/api/userApi";
 import type { DocumentData, Firestore, FirestoreError } from 'firebase/firestore';
 import { useParams } from 'react-router';
